Create Code.txt
You will need the European version of the game! (Just because this one has the most cheat codes available and is the best documented by me since I don't have the NTSC-U disc)
Simply right click the game on Dolphin's game list and click "Edit Config"
Add the cheat code with the $-comment beneath [GECKO] (add [GECKO] if it's not there).
Check the code on the list and have fun.
The new version consists of a cheat code because there's not enough static padding inside the .dol and it's easier to edit.
This includes the code by Dybbles to skip the tedious intro.